Add docs for awsiot.__init__ and greengrass_discovery

+++ make-docs.py
@@ -2,6 +2,7 @@
 
 import shutil
 import subprocess
+import sys
 
 try:
     import sphinx
@@ -9,10 +10,17 @@ except:
     exit('sphinx must be installed via pip')
 
 # Delete existing docs/, so that old no-longer-used files don't sit around forever
-shutil.rmtree('docs', ignore_errors=True)
+# This slows iteration down, so pass --cache to disable
+if not '--cache' in sys.argv:
+    shutil.rmtree('docs', ignore_errors=True)
 
 # Run sphinx
-subprocess.check_call(['sphinx-build', 'docsrc', 'docs'])
+subprocess.check_call([
+    'sphinx-build',
+    '-j', 'auto', # parallel build
+    '-W', # warnings as errors
+    'docsrc',
+    'docs'])
 
 # Create empty .nojekyll file, so that Github Pages shows plain old HTML
 with open('docs/.nojekyll', 'w'):
